kernel_10_26_1_18/jb-mr2-t40-r1-10.26.1.18-20131217-1558-foss/mydroid/external/webkit/LayoutTests/fast/events/touch/touchstart-prevent-default-prevents-click.html

<script src="../../js/resources/js-test-pre.js"></script>
<script src="./resources/touch-util.js"></script>
<style>
#testArea {
 width: 300px;
 height: 300px;
 background-color: blue;
}
</style>
<div ontouchstart="touchstart(); return false;" onclick="click()" id="testArea">
</div>
<div id="console"></div>
<script type="text/javascript">

if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
}

var touchStartCount = 0;
function touchstart() {
    console.log("touchstart");
    touchStartCount++;
    if (touchStartCount == 1)
        finishTest();
}

var clickCount = 0;
function click() {
    clickCount++;
    shouldBeTrue("false");
}

function finishTest() {
    shouldBe("touchStartCount", "1");
    shouldBe("clickCount", "0");
    if (window.layoutTestController) {
        isSuccessfullyParsed();
        layoutTestController.notifyDone();
    }
}

function testImmediate() {
    eventSender.clearTouchPoints();
    eventSender.addTouchPoint(100, 100);
    eventSender.touchStart();
    eventSender.touchEnd();
}

function testDelayed() {
    eventSender.clearTouchPoints();
    eventSender.addTouchPoint(100, 100);
    eventSender.touchStart();
}

function testDelayedResult() {
    eventSender.touchEnd();
}

var f = 0;
var functions = [
    [testImmediate, 0],
    // The following cases are commented out due to the flaky behavior
    // of the touch tester.
/*
    [testDelayed, 0],
    [testDelayedResult, 50],
    [testDelayed, 30],
    [testDelayedResult, 300],
*/
];

function runFunctions() {
    if (f == functions.length)
        return;

    var func = functions[f][0];
    var delay = functions[f][1];
    func();
    ++f;
    setTimeout(runFunctions, delay);
}

runTouchTest(function() {
    runFunctions();
}, "tap on the blue box");
</script>
